data_validation_automation_framework/requirements.txt
pandas
sqlalchemy
pyodbc
snowflake-connector-python
azure-storage-blob
pytest
openpyxl
jinja2
PyYAML 